Patch from James Montgomerie (jamie@montgomerie.net) to take MMAP into account
git-svn-id: svn+ssh://svn.abisource.com/svnroot/wv/trunk@20385 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ twips.c
@@ -49,7 +49,7 @@ float wvTwipsToMM(S16 twips)
 	{
 	float ret;
 	ret = ((float)twips)/TWIPS_PER_INCH;
-	ret = ret*25.0;
+	ret = ret*(float)25.0;
 	return(ret);
 	}
